fix login (add await to bcrypt.compare) and fix wrong password / user not found message.
There was an error upon creating a user. If the email was already in the database it still said "user created". To prevent this I added an "error" value to the response of the "takenEmail" check and then checked for that error in the createUser page.

For the Login page, I instead of returning an error on the api.js verifyUser method, I return the data.success, if false, the user couldn't login so I sent them an alert.